cli: allow overriding of bitcoin rpc host during auto-detection
This allows to use the existing auto-detection mechanism from the bitcoin
data dir, but to use a different host than the default 'localhost' that
is used at the moment. This is helpful in docker environments where
specter-desktop and bitcoind run in separate containers, since it is not
possible to connect to another container through localhost (except when
sharing the host network in the containers, which is not recommended).
By providing the hostname of the bitcoind container as `--bitcoin-rpc-host`
and mounting the bitcoind data volume into the specter-desktop container,
auto detection from the bitcoin.conf / cookie file works again without
having to supply rpc user, password, host, and port manually.