Enable macOS dock icon (#520)
Prior to this change, Specter desktop had no dock icon on macOS, only a
menu bar icon. This meant that users were forced to navigate with the
cursor to the menu bar icon and select 'Open Specter App' in order to
focus the Specter desktop window, resulting in a confusing and
cumbersome user experience.

The original commit of `specter_desktop.spec` in commit 2a799dde set the
macOS launch services `LSUIElement` key to `1` for unknown reasons.
This, along with what appears to be a default value of `False` for the
`LSBackgroundOnly` key have the effect of disabling the dock icon.

This change sets the value for both of these keys to to `False`,
enabling the dock icon to appear as expected. Experimentation showed
that eliminating one or both keys was not enough; they both must
explicitly be set to `False`.

See the Apple Information Property List Launch Services Keys developer
documentation for further details.

Fixes #455.