Improve tx-table and addresses-table render performance
By appending to tbody.innerHTML string, the browser throws away and
recreates the DOM nodes. This is slow and creates a lot of memory
pressure on the VM when done in a loop, as it was done here.

Now the DOMStrings are parsed into document fragments which are then
appended to the tbody. This way the existing elements in the tbody are
not thrown away and rebuilt each time a new item gets added.