Fix add keys (#1355)
* Fixes add key functionality for hot wallets

Earlier specter used to give an error 404 not found when trying to add
new keys to a hot wallet. This commit fixes this behaviour.

* Fixes add keys for hot wallets

The mnemonic value entered in the textbox while adding keys to existing
hot wallet was not used to add keys instead some random mnemonic was
used. This commit fixes this behaviour.