Update README.md (#2435)
Adding some clarity and consistency in how to provide user authentication details.

Both admin:password and admin:secret are used in curl examples, neither of which actually authenticate. 

Swapping them for user:password shows what the key-pair is meant to represent and then adding a note at the top to say what the default user and password key-value pair is allows for easy authentication.